adding forms (basket)

+++ app/models/CategoryRepository.scala
@@ -21,7 +21,7 @@ class CategoryRepository @Inject()(dbConfigProvider: DatabaseConfigProvider)(imp
 
   class CategoryTable(tag: Tag) extends Table[Category](tag, "category") {
 
-    def id = column[Int]("id", O.PrimaryKey, O.AutoInc)
+    def id = column[Long]("id", O.PrimaryKey, O.AutoInc)
 
     def name = column[String]("name")
